Refactor: all mentioned endpoints
